move chat_handler into a folder
